feat: 나라추천 budget 데이터타입 수정

+++ src/main/java/com/travel/rate/dto/travel/ReqTravelDTO.java
@@ -7,6 +7,6 @@ import lombok.Data;
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
 @Data
 public class ReqTravelDTO {
-    double budget;
+    Long budget;
     Long continent_id;
 }
